requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==5.0.2
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-crispy-forms==2.1
django-cruds-adminlte==0.0.16
django-flat-responsive==2.0
django-jazzmin==2.6.0
django-material-admin==1.8.6
django-redis==5.4.0
djangoajax==3.3
djangorestframework==3.14.0
idna==3.6
# mysqlclient==2.2.4
pillow==10.2.0
python-slugify==8.0.4
pytz==2024.1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.4
urllib3==2.2.0
